move `callback` module into `impl_`
